requirements.txt
django
python-telegram-bot
psycopg2-binary
python-dotenv